Add shared interface file to define shared interfaces
