+++ src/shared/interfaces/shared.interface.ts
@@ -1,0 +1,11 @@
+import { Bot } from './bot.interface';
+
+export interface ErrorResponse {
+  success: boolean;
+  message: string;
+  code?: string;
+  status?: number;
+  token: string | undefined;
+  bot?: Bot | undefined;
+  bots?: Bot[] | undefined;
+}
